requirements.txt
Flask
pymongo
Flask-PyMongo
pyjwt
gunicorn
redis
flask-testing
nose  